Readded last lines that haven been cut with previous convert method.
Basically i had a single config file for all these highlights, then i
separated the configs to their individual files using a vim macro. The
issue is that macro was not well tested and it accidentally deleted the
last lines of each of these files.
This commit fixes that by reading the missing lines.