feat: Added colorizer.lua
This new file contains the mix() function, which allows the user to mix
certain colors with either black or white by a certain factor, similar
to the mix() function found on css.

To use this function, you first have to require it in the file you want
to use it, for example with `local C = require("neopywal.colorizer")`.

With that, the mix() function will be avaliable for use, the syntax
follows a similar syntax from the css mix() function, namely:

mix([COLOR], "[BLACK | WHITE]", [FACTOR <= 255]).