refactor(util.lua): Improved overall code and added M.alpha().
- Added comments explaining what somethings do.
- Added hexToRgb() local function, which improves code readability by
simplifying the main functions like M.darken().
- Added M.alpha(), which allows to change the opacity value of any
color.

M.alpha() takes two parameters, "color" and "alpha".
the "color" parameter is self explanatory, while "alpha" is
an number betwheen 0 and 1, were 0 is fully transparent and 1 is fully
opaque.

It's important to note that, the current lastest version of Neovim
(0.10) does not support hex colors with more than six arguments (e.g
that this function does not add the extra two hex values at the end of
the hex string to change the opacity of colors. Instead it's uses clever
color mixing to create a new color that's just bright/dark enough to
give the illusion that it has some transparency, the effect is more
pronouced when using this function, for example, to set the background
color of a highlight group to a more "transparent" version, and setting
the foreground color of the highlight group to the original version of
the mixed color, giving the illusion of a fully opaque text on a
transparent background.