fix(init.lua): Fixed background color on Neovide with transparent option enabled.
Before this commit, if you set neopywal's transparent option to true and
opened Neovide instead of a terminal nvim, the background will be
rendered black because Neovide isn't able to render transparency
properly. This commit fixes that issue by disabling the transparent
option if the colorscheme is being used within Neovide.