refactor(editor): simplify TODO highlight
Before this commit, TODO comments were globally highlightted with the
background color being blue and the foreground color being the
background color. This behaviour was cofigured this way for an easier
time when configuring mini.hipatterns (and trouble.nvim it gets added).

However the TODO highlight is a builtin base highlight group, which
means it can be used by other plugins automatically, these plugins would
probably be expecting the TODO highlight to just have a simple
foreground color set, not what was configured previously by Neopywal.
This could cause misbehaviour for those plugins, that's why this
highlight group has been changed to match the expectations.