feat!: colorscheme compiler
This commit implements an compiler for the Neopywal colorscheme.
The code was mostly yanked from the catppuccin Neovim theme
(github.com/catppuccin/nvim). With some tweaks made to fit the code
structure of Neopywal and with some additional features (mainly for
debugging).

All the compiled files can be found on the nvim cache directory inside
the neopywal directory. The colorscheme get's recompiled automatically
when any change has been made in the plugin configuration
(with `require("neopywal").setup()`).

The name for the cache files is based on the sum of all hexadecimal
colors that get exported by the `require("neopywal").get_colors()`
function. The colors are first converted into a set of three RGB
numbers with the `require("neopywal.util").hexToRgb` function. And
everything get's summed together and the final resulting number will be
used as the file name for the compiled cache file.

In that way, if the user changes their pywal theme, the hexadecimal
values will be different than the previous compiled Neopywal cache,
therefore the sum of all hexadecimal colors will also be different,
which will result in a different file name that get's used for the
compiled cache.