refactor(compiler): "neopywal" as cached filename
Before this commit, the filename used for the neopywal compiler was
calculated using the sum of all colors within the Neopywal color
palette, that meant that, everytime the user changed their pywal
colorscheme, a new cache file was going to be created by the compiler
with a unique name based on that sum. This behaviour has been changed,
now the compiler always looks for the same filename defined with
`require("neopywal").compiled_filename`, and the color palette sum is
instead appended to the hash used on the `cached` file. In this way
Neopywal can tell if the colorscheme needs to be recompiled because of
changes in the color palette without having to create a million
different cache files (which i guess saves HDD space xppp).