perf: limit sum_colors() table
This commit improves the `sum_colors()` function by reducing the total
count of HEX colors from the palette exportted by `M.get_colors()`. By
reducing the total amount of colors the function sums the total hash
accuracy also goes down, that's because with all colors main colors
being used, any change made to any of them would automatically trigger
the recompilation of the colorscheme. However with less colors being
detected, the higher is the change of some change on a non-supported
color not being detected and therefore not triggering the recompilation
of the colorscheme.

Some colors have been disabled by default because they are not part of
the main colorscheme, so changes in them won't affect anything in the
colorscheme, which means no recompilation needed. And even if the user
changed one of the disabled colors in their configuration, that would
mean a change in the `user_conf` variable inside `M.setup()`, which in
turn would also mean a different calculated cached hash, since that
value depends on the user configuration table to be calculated.

The only problem i see with that new approach for `sum_colors()` is if
the user changes one of the non-tracked colors directly in the auto
generated pywal color template file. But that is very unlikely to
happen and, even if it did, it can be easily mitigated by a future
option on `M.setup()` that allows the direct modificatin of the
behaviour of `sum_colors()`.

This commit in a way also fixes a bug in `sum_colors()`, before this
commit, if the user added extra colors in the color table for
`M.get_colors()`, the sum of **all** of them would also be calculated by
`sum_colors()`. But now `sum_color()` has been limited to only calculate
the sum of the colors defined inside the function.