refactor: "default_" option handling
This commit changes the way the "default_fileformats" and
"default_plugins" options are handled.

Before this commit, the way this these options modified the
fileformats/plugins tables was the following:

If they were set to false in the USER configuration, their respective
option in the "default_options" table was going to be modified to be
empty. Then the M.options table was going to be evaluatued, which would
concatenate the "default_options" table with an empty
fileformats/plugins table, which would be overwritten by the user config
if the user specified the contents of them manually on their setup()
function.

E.g.: if the user set default_fileformats to false, the
"fileformats" table inside the default_options table was going to be
modified to be an empty table. Then M.options would concatenate
default_options.fileformats = {} with the user configuration, and if the
user manually set a fileformats table on their setup() function. Their
configuration table would be the one present in M.options which in turn
would be exportted by Neopywal.

This behaviour has been changed. Now M.options is evaluatued BEFORE
anything is done with the "default_" option tables. This change was
mainly made because, for example, if Neopywal changed
"default_fileformats" to false in the default options, that change
wouldn't have been evaluatted when calling setup(). That's because the
"default_fileformats" and/or "default_plugins" tables were only being
checked by changes made by the user, not changes made in the
"default_options" table.

Obviously this does not change the user experience. It's just a change
to be able to better handle the "M.options" table internally.