feat: treesitter disables "default_fileformats"
This commit changes the behavior of the "default_fileformats" option.
Now it will be turned off by default if the treesitter plugin option is
enabled (unless the user manually enables the option).

This change was mainly made of better performance when Neopywal is
loading it's compiled cache. Since there's no need to load all of the
fileformats highlight groups if treesitter is in use.