refactor: local default_options table
The default_options table was previously in use in a old borrowed code
inside the compiler.lua "map_highlights()" function. So because of that
it had to be globally exported inside the "M" table.

The problem is that the borrowed code really didn't serve any purpose,
so it was removed, therefore making the "default_options" table globally
exportable is no longer needed.

The reason it's best practice to make such table only locally available
is because, when it get's exported, the user can externally modify it
with 'require("neopywal").default_options' despite it being a table that
really shouldn't ever be modified as it contains all the bare options
structured in the intended way needed for Neopywal to function.