fix(compiler): malformed string formatting
Apparently the "string.format" function doesn't properly concatenate
multiple string paramaters at once, because it tries to replace
something "%s" in the first string parameter with the following strings.

This caused an error on one of the print functions inside the
compiler.lua file, specifically the print statement that tells something
went wrong with the compilation.