refactor(kinds): resolve highlight links
Refactored the "kinds.lua" so that it no longer export links to
syntax highlight groups for each kind. Instead it now has the colors for
all highlight groups set within the module itself.

Using this approach has two benefits. The first is that, because no
links to highlight groups were used, none of the styles applied to these
groups is carried over. Making the exported "kinds" table cleaner.

The second benefit is an addition parameter that was added to the
"M.get()" function. This parameter allows to apply the contents of a
table directly to all groups exported by the kinds module.